Fixes #582 "Exporter deletes items on 'Extract Only'"
The exporter would try to insert an item, and re-insert back the remainder if there is no space. However, for storages that use "extract only" the exporter would not be able to re-insert back the remainder, which leads to item loss.

To prevent this you can now run a simulate check on extractItem so the exporter can first check *if* the item can be inserted, and be sure there is no remainder.