Fix connections conductivity (#1089)
* Check neighboring INetworkNode blocks for canConduct before connecting

* Add getConnectableConditions() check back in and handle canConduct
separately for INetworkNode connections

* Generalize the post-getConnectableConditions test for TileMultipartNode to
include all INetworkNodes and return canConduct instead of the funuctional
equivalent of TileMultipartNode's canConduct method.

* Add a hasConnectableConditions method to the public API which performs the
getConnectableConditions stream test on the provided TileEntity.  Some
Mods (EIO) are not Java 8 and can't use lambda functions to test for themselves.
Use hasConnectableConditions method in BlockCable for testing in
hasConnectionWith.