Added more Item support framework, refactored sanitation
Found one case in which to never ingore NBT tags, which is when dealing with items inside of a grid. Otherwise items with a tag being ignored will become exactly the same as the first equal item found. Added a NO_FORESTRY_SANTIZE flag to mark these cases, and marked the insert method for StorageDIsks. This prevents the (Forestry) item from being converted into other "equal" items, but leaves GUI bugs.
Unsure if I should continue trying to flag non-sanitation instances, or to instead mark methods that should sanitize.