More code cleanup, organized imports
All buttons set the GEN and IS_ANALYZED flags for sanitation. Those tags are the most common tags most would want to ignore with autonamation in Forestry. No buttons are loaded if Forestry is not present. Decided the filter item would be best for any more complicated Forestry NBTTag filtering if desired but the Filter GUI is already quite full.