Changed item quantity of a craftable item to 0 (#1613)
* Changed item quantity of a craftable item to 0

Whenever a system has none of a specific item but the ability to craft it, the grid saves the quantity as 1. Quantity sorting would make use of this and sort it alphabetically together with other quantity 1 items. This change should make quantity sorting think craftable items have a quantity of 0, making all of them appear behind itemstacks with the quantity 1 (or higher).

* Moving code logic to GridStackItem

* Update GridSortingQuantity.java

* Changed stack quantity logic

Craftable itemstacks remain stored as "1 item" internally. Other classes will be given the information that the stack is empty, though.

* Update CHANGELOG.md