Allow crafters to be daisy-chained
In crafting situations that require more than the 9 blocking patterns
that one crafter allows, the only solution is to duplicate the setup
to allow a second crafter.  This CL allows crafters to be
daisy-chained to provide a solution to this issue.

Also, since unlimited crafters can now be placed on a single
inventory, this eliminates the need to have multiple intermediate
inventories which then pipe into the desired inventory.  This also
improves the crafting manager UI as now those patterns can all be
listed under one "Furnace" instead of (potentially multiple) "Chest"s.