Potential fix NPE in ItemGridView::postChange occurring on client.
This appears to happen when the contents of the grid previously
contained items that were also ingredients for an autocrafting recipe,
and the autocrafting recipe has been created.

Based on the null check in the block of code after the one edited, it
seems to be a thing that the map can potentially return a null result;
however, the "update other id" block does not test for null, simply
calling `updateOtherId`.

This commit adds a test for a potential null result from the map.

While this might just be addressing the side effect of a larger issue
associated with auto-crafting (i.e., the expectation that the IGridStack
wouldn't be a null result), it hopefully reduces the number of annyoing
client-side crashes.