Move GridStack* quantity manipulation out of GridView classes
This follows the Law of Demeter a little better. It also removes the
need for different postChange functions (because they are functionally
identical except for the specifics of how the stack quantity is edited),
which makes the GridView subclasses redundant, so they can be merged
back into the base class.