Speed up chunk loading when many similar items in RS
When there are similar items (same type, different enchantments and
damage) in an RS system, it can sometimes take a very long time to load.

This should take a stab at fixing that, with the caveat that it
potentially reduces the flexibility of `IStackList#get` and
`IStackList#getEntry` (now forces at least COMPARE_NBT, though as far as
I can tell there was never a time when this wasn't the case already).